% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cp.R
\name{plot_cp}
\alias{plot_cp}
\title{Contraceptive Relavence Rate Over Time}
\usage{
plot_cp(dat, est, iso_code, CI = NULL)
}
\arguments{
\item{dat}{A data frame containing observed contraceptive use data with columns:
\code{iso} (country code), \code{year} (reference year), and \code{cp} (contraceptive use modern \%).}

\item{est}{A data frame containing estimated contraceptive use data with columns:
\code{iso}, \code{Year}, \code{Median}, \code{L95}, \code{U95}, \code{L80}, \code{U80}.}

\item{iso_code}{Numeric Value for a country code}

\item{CI}{The confidence interval for the graphs, either 95\%, 80\% or other.}
}
\value{
A ggplot object displaying the contraceptive use over time.
}
\description{
Contraceptive Relavence Rate Over Time
}
